fixcomment_ss2

+++ ss2_spring_controller/exercise/caculator/src/main/java/com/example/caculator/service/ICalculatorService.java
@@ -7,5 +7,5 @@ public interface ICalculatorService {
 
     int multiplication(int numberOne, int numberTwo);
 
-    double division(int numberOne, int numberTwo);
+    double division(double numberOne, double numberTwo);
 }
