+++ ss2_spring_controller/exercise/caculator/src/main/java/com/example/caculator/service/impl/CalculatorService.java
@@ -21,7 +21,7 @@ public class CalculatorService implements ICalculatorService {
     }
 
     @Override
-    public double division(int numberOne, int numberTwo) {
+    public double division(double numberOne, double numberTwo) {
         return numberOne / numberTwo;
     }
 }
